Fix bug with packet receive.
Notch should study more about network programming, the client was
sending 2 different packets, 1 big 1 byte with the packet id and another
with the actual packet content. How to waste bandwidth with unneeded
overhead.

Now the cd_ReadCallback waits for the buffer to be larger than 1, so it
doesn't miss packets.